backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
flask-cors==5.0.0
flask-socketio==5.4.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
twilio==9.4.3
gunicorn==23.0.0
eventlet==0.38.2
pytest==8.3.3
pytest-flask==1.3.0
